add cross reference file, use as base, change postinstall to update script for manual update if ever required

+++ updatejson.sh
@@ -4,8 +4,11 @@
 fn=windowsZones.xml
 #  download unicode.org supported  list of windows/iana timezone mappings
 curl -sL https://raw.githubusercontent.com/unicode-org/cldr/master/common/supplemental/$fn >$fn
-# convert to json
-node node_modules/xml-js/bin/cli.js $fn >$(basename $fn .xml).json
-# erase xml file
-rm $fn
-node cvt-to-json.js
+res=$?
+if [ $eres -eq 0 ]; then
+	# convert to json
+	node node_modules/xml-js/bin/cli.js $fn >$(basename $fn .xml).json
+	# erase xml file
+	rm $fn
+	node cvt-to-json.js
+fi
